Refactor Test classes to separate files
