+++ tests/test_answers.py
@@ -348,14 +348,3 @@ class TestUsers(unittest.TestCase):
         reply = json.loads(response.data.decode())
 
         self.assertEqual(reply['message'], 'Barna is logged in.')
-
-
-class ModelsTests(unittest.TestCase):
-    def test_user_model(self):
-        user = User('1', 'Barna', 'barna@gmail.com', '12345')
-
-    def test_question_model(self):
-        question = Question('1', '2', 'what is coding?')
-
-    def test_answer_model(self):
-        answer = Answer('1', '2', 'coding is obulamu')
